+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## [1.1.3](https://github.com/Nyakaze/Cordi/compare/1.1.2...1.1.3) (06-02-2025)
+## [1.1.4](https://github.com/Nyakaze/Cordi/compare/1.1.3...1.1.4) (07-02-2026)
+
+### General
+ - **Added:** Experimental Warning for Discord Commands.
+ - **Added:** Tooltip for Chat Mappings.
+ - **Added:** Stats to the Config Window.
+ - **Fixed:** Tooltips not beeing in the top right corner.
+ - **Updated:** Discord Bot Button to be disabled when the bot is starting or stopping.
+ - **Changed:** Cordi Peep window to Peeper window.
+ - **Changed:** Discord Bot Token input to hide the token.
+
+
+## [1.1.3](https://github.com/Nyakaze/Cordi/compare/1.1.2...1.1.3) (06-02-2026)
 
 ### Webhook Service
  - **Fixed:** Webhooks sometimes throwing Errors. Should now reliably send messages.